run automated tests for different devices in parallel
