+++ shell_tests/automation_tests/test_connectivity.py
@@ -30,9 +30,8 @@ class TestConnectivity(BaseTestCase):
 
         with DutHandler(
                 cs_handler,
-                self.resource_handler.reservation_id,
                 self.shell_conf.dut_shell_path,
-                self.shell_conf.dut_dependencies_path if self.shell_conf.dependencies_path else None,
+                self.resource_handler.reservation_id,
                 self.logger) as dut_handler:
 
             res_info = self.resource_handler.get_details()
